% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tldrDocs-package.R
\docType{package}
\name{tldrDocs}
\alias{tldrDocs-package}
\alias{tldrDocs}
\alias{package-tldrDocs}
\title{tldrDocs: tldr-style documentation for base R packages}
\description{
A collection of tldr-style documentation for commonly-used objects exported in
the base distribution of R.
}
\seealso{
Useful links:
\itemize{
\item \url{https://github.com/jamesotto852/tldrDocs/}
\item \url{https://github.com/jamesotto852/tldrPages/}
}
}
\author{
\strong{Maintainer}: James Otto \email{jamesotto852@gmail.com} (\href{https://orcid.org/0000-0002-0665-2515}{ORCID})

Authors:
\itemize{
  \item David Kahle \email{david@kahle.io} (\href{https://orcid.org/0000-0002-9999-1558}{ORCID})
}

}
